refactor: move imageref.Version helper to shared/utils
